Move Send to the token domain
