+++ src/locales/en/common.json
@@ -3,6 +3,7 @@
   " is out. ": " is out. ",
   " to deposit LSK.": " to deposit LSK.",
   "\"{{username}}\" is already taken.": "\"{{username}}\" is already taken.",
+  "64 bytes left": "64 bytes left",
   "A bit more. Make sure to type at least 3 characters.": "A bit more. Make sure to type at least 3 characters.",
   "A great way to start is to top up your account with some {{value}}.": "A great way to start is to top up your account with some {{value}}.",
   "About": "About",
@@ -16,6 +17,7 @@
   "Active delegates": "Active delegates",
   "Add": "Add",
   "Add account": "Add account",
+  "Add address to bookmarks": "Add address to bookmarks",
   "Add another account": "Add another account",
   "Add bookmark": "Add bookmark",
   "Add new": "Add new",
@@ -66,6 +68,7 @@
   "Blocks forged": "Blocks forged",
   "Blocks overview": "Blocks overview",
   "Bookmarks": "Bookmarks",
+  "Bytes counter": "Bytes counter",
   "Cancel": "Cancel",
   "Caution! This delegate was punished on {{punishmentStartDate}}. There is approximately {{daysLeft}} days remaining before the punishment ends.": "Caution! This delegate was punished on {{punishmentStartDate}}. There is approximately {{daysLeft}} days remaining before the punishment ends.",
   "Caution! You are about to send the majority of your balance": "Caution! You are about to send the majority of your balance",
@@ -139,6 +142,7 @@
   "Duplicate public keys detected.": "Duplicate public keys detected.",
   "Edit": "Edit",
   "Edit bookmark": "Edit bookmark",
+  "Edit transaction": "Edit transaction",
   "Edit vote": "Edit vote",
   "Either change the optional member to mandatory or define more optional members.": "Either change the optional member to mandatory or define more optional members.",
   "Either change the optional member to mandatory or reduce the number of signatures.": "Either change the optional member to mandatory or reduce the number of signatures.",
@@ -195,12 +199,14 @@
   "Input transaction data": "Input transaction data",
   "Input view": "Input view",
   "Insert a vote amount for this delegate. Your new vote will be added to the voting queue.": "Insert a vote amount for this delegate. Your new vote will be added to the voting queue.",
+  "Insert address or bookmark name": "Insert address or bookmark name",
   "Insert entire signed message": "Insert entire signed message",
   "Insert label": "Insert label",
   "Insert message": "Insert message",
   "Insert public address": "Insert public address",
   "Insert public key": "Insert public key",
   "Insert signature": "Insert signature",
+  "Insert transaction amount": "Insert transaction amount",
   "Insert vote amount": "Insert vote amount",
   "Inside round": "Inside round",
   "Install update": "Install update",
@@ -218,6 +224,7 @@
   "Lisk": "Lisk",
   "Lisk Terms of Use": "Lisk Terms of Use",
   "Lisk Website": "Lisk Website",
+  "Lisk counts your message in bytes, so keep in mind\n                that the length of your message may vary in different languages.\n                Different characters may consume a varying amount of bytes.": "Lisk counts your message in bytes, so keep in mind\n                that the length of your message may vary in different languages.\n                Different characters may consume a varying amount of bytes.",
   "Lisk has now been enhanced even further": "Lisk has now been enhanced even further",
   "Lisk wallet": "Lisk wallet",
   "Lisk {{version}}": "Lisk {{version}}",
@@ -239,6 +246,7 @@
   "Medium": "Medium",
   "Members": "Members",
   "Message": "Message",
+  "Message (optional)": "Message (optional)",
   "Min": "Min",
   "Minimize": "Minimize",
   "Missed slot": "Missed slot",
@@ -312,6 +320,7 @@
   "Privacy policy": "Privacy policy",
   "Provide Feedback": "Provide Feedback",
   "Provide a correct amount of {{token}}": "Provide a correct amount of {{token}}",
+  "Provide a correct wallet address or the name of a bookmarked account": "Provide a correct wallet address or the name of a bookmarked account",
   "Provide a signature for a transaction which belongs to a multisignature account.": "Provide a signature for a transaction which belongs to a multisignature account.",
   "Provided amount is higher than your current balance.": "Provided amount is higher than your current balance.",
   "Provided amount will result in a wallet with less than the minimum balance.": "Provided amount will result in a wallet with less than the minimum balance.",
@@ -373,7 +382,9 @@
   "Select the applicable mode.": "Select the applicable mode.",
   "Send": "Send",
   "Send a reclaim transaction": "Send a reclaim transaction",
+  "Send maximum amount": "Send maximum amount",
   "Send using second passphrase right away": "Send using second passphrase right away",
+  "Send {{amount}} {{token}}": "Send {{amount}} {{token}}",
   "Send {{token}}": "Send {{token}}",
   "Send {{token}} here": "Send {{token}} here",
   "Sender": "Sender",
@@ -495,6 +506,7 @@
   "Whoops, that page is gone.": "Whoops, that page is gone.",
   "Window": "Window",
   "Would you like to download it now?": "Would you like to download it now?",
+  "Write message": "Write message",
   "You are about to send your entire balance": "You are about to send your entire balance",
   "You can download or copy the transaction and share it with other members.": "You can download or copy the transaction and share it with other members.",
   "You can learn more": "You can learn more",
@@ -544,6 +556,7 @@
   "{{VOTE_LIMIT}} votes available for your account": "{{VOTE_LIMIT}} votes available for your account",
   "{{acctToken}} balance": "{{acctToken}} balance",
   "{{amount}} {{token}} Received": "{{amount}} {{token}} Received",
+  "{{length}} bytes left": "{{length}} bytes left",
   "{{passphraseName}} paper wallet": "{{passphraseName}} paper wallet",
   "{{transactions}} transactions": "{{transactions}} transactions"
 }
