+++ src/modules/token/fungible/components/SendForm/form.css
