+++ src/modules/token/fungible/components/SendForm/useMessageField.js
