+++ src/modules/token/fungible/components/SendStatus/status.css
