+++ src/modules/token/fungible/components/SendSummary/summary.js
