+++ src/modules/token/fungible/components/SendView/index.js
@@ -6,9 +6,9 @@ import MultiStep from 'src/modules/common/components/OldMultiStep';
 import { parseSearchParams } from 'src/utils/searchParams';
 import TxSignatureCollector from '@transaction/components/TxSignatureCollector';
 import Dialog from 'src/theme/dialog/dialog';
-import Form from './form';
-import Summary from './summary';
-import Status from './status';
+import Form from '../SendForm';
+import Summary from '../SendSummary';
+import Status from '../SendStatus';
 import styles from './send.css';
 
 const Send = ({ history }) => {
