+++ src/modules/token/fungible/components/SendView/index.test.js
