+++ src/modules/token/fungible/components/SendView/send.css
